+++ delivery-market/src/main/java/com/ddangme/dm/controller/good/api/CartApiController.java
@@ -1,4 +1,4 @@
-package com.ddangme.dm.controller.member.api;
+package com.ddangme.dm.controller.good.api;
 
 
 import com.ddangme.dm.dto.cart.request.CartChangeCheckRequest;
@@ -7,10 +7,7 @@ import com.ddangme.dm.dto.cart.request.CartRequest;
 import com.ddangme.dm.dto.cart.response.CartListResponse;
 import com.ddangme.dm.dto.cart.response.CartResponse;
 import com.ddangme.dm.dto.member.MemberPrincipal;
-import com.ddangme.dm.dto.pick.PickedGoodResponse;
-import com.ddangme.dm.service.FileService;
 import com.ddangme.dm.service.good.CartService;
-import com.ddangme.dm.service.good.PickService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
@@ -24,34 +21,11 @@ import java.util.List;
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/api")
-public class MemberGoodApiController {
+public class CartApiController {
 
-    private final PickService pickService;
     private final CartService cartService;
-    private final FileService fileService;
 
-    @GetMapping("/my-page/pick/list")
-    public ResponseEntity<List<PickedGoodResponse>> pickList(@AuthenticationPrincipal MemberPrincipal principal) {
-        return ResponseEntity.ok(pickService.findPickedGood(principal.getId()));
-    }
 
-    @DeleteMapping("/goods/pick/{goodId}")
-    public ResponseEntity<Void> deletePick(@AuthenticationPrincipal MemberPrincipal principal, @PathVariable Long goodId) {
-        pickService.deletePick(principal.getId(), goodId);
-        return ResponseEntity.ok().build();
-    }
-
-    @PostMapping("/goods/pick/{goodId}")
-    public ResponseEntity<Boolean> pick(@PathVariable Long goodId,
-                                        @AuthenticationPrincipal MemberPrincipal principal) {
-        return ResponseEntity.ok().body(pickService.pick(goodId, principal.getId()));
-    }
-
-    @GetMapping("/goods/find/pick/{goodId}")
-    public ResponseEntity<Boolean> findPick(@PathVariable Long goodId,
-                                            @AuthenticationPrincipal MemberPrincipal principal) {
-        return ResponseEntity.ok().body(pickService.findPick(goodId, principal.getId()));
-    }
 
     @PostMapping("/goods/cart")
     public ResponseEntity<CartResponse> cart(@RequestBody List<CartRequest> requests,
